Safari-611.1.10/LayoutTests/fast/dom/htmlcollection-length-after-item-2.html

<!DOCTYPE html>
<script src="../../resources/js-test-pre.js"></script>
<script>

description("This tests accessing the length after accessing the first item in an empty HTMLCollection doesn't cache a wrong length.");

var container = document.createElement('div');
var span = document.createElement('span');
var children;
shouldBe("children = container.children; children.item(0)", "null");
shouldBe("children.length", "0");

</script>
<script src="../../resources/js-test-post.js"></script>
